LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigate-event/intercept-multiple-times-expected.txt


FAIL navigation.navigate() returns a finished promise that awaits all promises if event.intercept() is called multiple times assert_equals: expected "#1" but got ""
